requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.14.0
pymongo==3.11.2
requests==2.25.0
six==1.15.0
urllib3==1.26.2
